requirements.txt
ultralytics==8.3.4
opencv-python
numpy
torch
torchvision
scikit-learn
PyYAML
tqdm